+++ tests/TestModels/SalesOrderWithAfterTransitionHook.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Asantibanez\LaravelEloquentStateMachines\Tests\TestModels;
+
+use Asantibanez\LaravelEloquentStateMachines\Tests\TestStateMachines\SalesOrders\StatusWithAfterTransitionHookStateMachine;
+use Asantibanez\LaravelEloquentStateMachines\Traits\HasStateMachines;
+use Illuminate\Database\Eloquent\Model;
+
+class SalesOrderWithAfterTransitionHook extends Model
+{
+    use HasStateMachines;
+
+    protected $table = 'sales_orders';
+
+    protected $guarded = [];
+
+    public $stateMachines = [
+        'status' => StatusWithAfterTransitionHookStateMachine::class,
+    ];
+}