+++ tests/TestStateMachines/SalesOrders/StatusWithAfterTransitionHookStateMachine.php
@@ -1,0 +1,48 @@
+<?php
+
+
+namespace Asantibanez\LaravelEloquentStateMachines\Tests\TestStateMachines\SalesOrders;
+
+
+use Asantibanez\LaravelEloquentStateMachines\StateMachines\StateMachine;
+use Asantibanez\LaravelEloquentStateMachines\Tests\TestJobs\AfterTransitionJob;
+
+class StatusWithAfterTransitionHookStateMachine extends StateMachine
+{
+    public function recordHistory(): bool
+    {
+        return true;
+    }
+
+    public function transitions(): array
+    {
+        return [
+            'pending' => ['approved'],
+            'approved' => ['processed'],
+        ];
+    }
+
+    public function defaultState(): ?string
+    {
+        return 'pending';
+    }
+
+    public function afterTransitionHooks(): array
+    {
+        return [
+            'pending' => [
+                function($to, $model) {
+                    $model->total = 200;
+                    $model->save();
+                },
+                function($to, $model) {
+                    $model->notes = 'after';
+                    $model->save();
+                },
+                function ($to, $model) {
+                    AfterTransitionJob::dispatch();
+                },
+            ]
+        ];
+    }
+}